fix/issue-384: extract EdgePanelOverlay; converge IDLE_DIM_MS; document axisLockDp; fix barrel
- EdgePanelOverlay: shared pan handler, worklets, retraction logic;
  params: zone, onCommit, sheetHeightFraction, commitPredicate, children
- ControlCenterOverlay: thin wrapper (commitForPanel, 0.55, CC body)
- NotificationCenterOverlay: thin wrapper (commitForNC, 0.65, NC body)
- currentT seed converged to 0 in both (was e.absoluteY in CC — latent bug)
- IDLE_DIM_MS moved to gestureConfig.ts; HomeIndicator + AssistiveTouch import it
- axisLockDp multipliers (×8, ×1, ×2) documented with justification comments
- barrel: exports AssistiveTouch, CC/NC/Spotlight overlays, EdgePanelOverlay
