+++ src/components/ControlCenterOverlay.tsx
@@ -1,18 +1,7 @@
-import React, { useEffect } from 'react';
-import { View, Text, StyleSheet, Dimensions } from 'react-native';
-import { useSafeAreaInsets } from 'react-native-safe-area-context';
-import Animated, {
-  useSharedValue,
-  useAnimatedStyle,
-  runOnJS,
-} from 'react-native-reanimated';
-import { Gesture, GestureDetector } from 'react-native-gesture-handler';
-import { gestureConfig } from '../utils/gestureConfig';
-import { useVelocityBuffer, pushSample, sampledVelocity } from '../utils/gestureVelocity';
+import React from 'react';
+import { View, Text, StyleSheet } from 'react-native';
 import { commitForPanel } from '../utils/gestureMachine';
-import { settle, useGestureReduceMotion } from '../utils/useGestureReduceMotion';
-
-const { height: SCREEN_HEIGHT } = Dimensions.get('window');
+import { EdgePanelOverlay } from './EdgePanelOverlay';
 
 interface Zone {
   top: number;
@@ -27,157 +16,25 @@ interface Props {
 }
 
 export function ControlCenterOverlay({ zone, onCommit }: Props) {
-  const insets = useSafeAreaInsets();
-  const reduceMotion = useGestureReduceMotion();
-  const reduceMotionShared = useSharedValue(reduceMotion);
-  useEffect(() => {
-    reduceMotionShared.value = reduceMotion;
-    // Shared values are stable refs; only respond to reduceMotion changes
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [reduceMotion]);
-
-  const panelProgress = useSharedValue(0);
-  const buf = useVelocityBuffer();
-  const startedInZone = useSharedValue(false);
-  // Wall-clock timestamp updated per gesture event
-  const currentT = useSharedValue(0);
-
-  const pan = Gesture.Pan()
-    .onBegin((e) => {
-      'worklet';
-      buf.value = [];
-      startedInZone.value =
-        e.absoluteX >= zone.left &&
-        e.absoluteX <= zone.right &&
-        e.absoluteY >= zone.top &&
-        e.absoluteY <= zone.bottom;
-      currentT.value = e.absoluteY; // approximate; overwritten in onUpdate
-    })
-    .onUpdate((e) => {
-      'worklet';
-      if (!startedInZone.value) return;
-      currentT.value = Date.now();
-      pushSample(buf.value, e.translationX, e.translationY, currentT.value);
-      const dy = Math.max(0, e.translationY);
-      panelProgress.value = Math.max(0, Math.min(1, dy / gestureConfig.panelTravelDp));
-    })
-    .onEnd((e) => {
-      'worklet';
-      if (!startedInZone.value) {
-        panelProgress.value = settle(0, 'fastSettle', reduceMotionShared.value);
-        return;
-      }
-      currentT.value = Date.now();
-      pushSample(buf.value, e.translationX, e.translationY, currentT.value);
-      const { vy } = sampledVelocity(buf.value, currentT.value);
-      const progress = panelProgress.value;
-      const reason = commitForPanel({ progress, velocity: vy, holdMs: 0 });
-      if (reason !== 'none') {
-        // Hand off to the real screen: retract the preview so it doesn't
-        // linger underneath the transparent modal and block the home screen
-        // when the user returns.
-        panelProgress.value = settle(0, 'fastSettle', reduceMotionShared.value);
-        runOnJS(onCommit)();
-      } else {
-        panelProgress.value = settle(0, 'fastSettle', reduceMotionShared.value);
-      }
-    });
-
-  const sheetStyle = useAnimatedStyle(() => {
-    'worklet';
-    const translateY = -SCREEN_HEIGHT * (1 - panelProgress.value);
-    // When fully retracted, drop the view out of the render tree so its
-    // layout-based hit area cannot absorb taps on the home content behind it.
-    const display = panelProgress.value <= 0.001 ? 'none' : 'flex';
-    return {
-      transform: [{ translateY }],
-      display,
-    };
-  });
-
-  const backdropStyle = useAnimatedStyle(() => {
-    'worklet';
-    const opacity = panelProgress.value * 0.5;
-    const display = panelProgress.value <= 0.001 ? 'none' : 'flex';
-    return {
-      opacity,
-      display,
-    };
-  });
-
-  const zoneWidth = zone.right - zone.left;
-  const zoneHeight = zone.bottom - zone.top;
-
   return (
-    <>
-      {/* Dark backdrop */}
-      <Animated.View
-        style={[StyleSheet.absoluteFill, styles.backdrop, backdropStyle]}
-        pointerEvents="none"
-      />
-
-      {/* Sliding preview sheet — visual only, hidden from screen readers */}
-      <Animated.View
-        style={[styles.sheet, { paddingTop: insets.top + 12 }, sheetStyle]}
-        pointerEvents="none"
-        accessible={false}
-        importantForAccessibility="no-hide-descendants"
-      >
-        <View style={styles.handle} />
-        <Text style={styles.title}>Control Center</Text>
-        <View style={styles.tileRow}>
-          <View style={styles.tile} />
-          <View style={styles.tile} />
-          <View style={styles.tile} />
-          <View style={styles.tile} />
-        </View>
-      </Animated.View>
-
-      {/* Activation zone — intercepts touches in the top-right corner; hidden from TalkBack */}
-      <View
-        style={[
-          styles.activationZone,
-          {
-            top: zone.top,
-            left: zone.left,
-            width: zoneWidth,
-            height: zoneHeight,
-          },
-        ]}
-        pointerEvents="auto"
-        accessible={false}
-        importantForAccessibility="no-hide-descendants"
-      >
-        <GestureDetector gesture={pan}>
-          <View style={StyleSheet.absoluteFill} />
-        </GestureDetector>
+    <EdgePanelOverlay
+      zone={zone}
+      onCommit={onCommit}
+      sheetHeightFraction={0.55}
+      commitPredicate={commitForPanel}
+    >
+      <Text style={styles.title}>Control Center</Text>
+      <View style={styles.tileRow}>
+        <View style={styles.tile} />
+        <View style={styles.tile} />
+        <View style={styles.tile} />
+        <View style={styles.tile} />
       </View>
-    </>
+    </EdgePanelOverlay>
   );
 }
 
 const styles = StyleSheet.create({
-  backdrop: {
-    backgroundColor: '#000',
-  },
-  sheet: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    right: 0,
-    height: SCREEN_HEIGHT * 0.55,
-    backgroundColor: 'rgba(28,28,30,0.92)',
-    borderBottomLeftRadius: 20,
-    borderBottomRightRadius: 20,
-    alignItems: 'center',
-  },
-  handle: {
-    width: 36,
-    height: 4,
-    borderRadius: 2,
-    backgroundColor: 'rgba(255,255,255,0.35)',
-    marginBottom: 16,
-  },
   title: {
     color: '#fff',
     fontSize: 16,
@@ -193,8 +50,5 @@ const styles = StyleSheet.create({
     height: 56,
     borderRadius: 14,
     backgroundColor: 'rgba(255,255,255,0.12)',
-  },
-  activationZone: {
-    position: 'absolute',
   },
 });
