+++ src/components/index.ts
@@ -26,3 +26,8 @@ export { GestureHost, useGestureHost, useOptionalGestureHost } from './GestureHo
 export type { GestureHostMode } from './GestureHost';
 export { BackEdgeSwipe } from './BackEdgeSwipe';
 export { QuickSwitchHomeBar } from './QuickSwitchHomeBar';
+export { AssistiveTouch } from './AssistiveTouch';
+export { ControlCenterOverlay } from './ControlCenterOverlay';
+export { NotificationCenterOverlay } from './NotificationCenterOverlay';
+export { SpotlightReveal } from './SpotlightReveal';
+export { EdgePanelOverlay } from './EdgePanelOverlay';
